refactor(sign): Change files' names
